+++ src/test/groovy/ru/vyarus/dropwizard/guice/lifecycle/ShutdownEventTest.groovy
@@ -9,6 +9,7 @@ import org.junit.runners.model.Statement
 import ru.vyarus.dropwizard.guice.GuiceBundle
 import ru.vyarus.dropwizard.guice.module.lifecycle.GuiceyLifecycleAdapter
 import ru.vyarus.dropwizard.guice.module.lifecycle.event.jersey.ApplicationShotdownEvent
+import ru.vyarus.dropwizard.guice.module.lifecycle.event.jersey.ApplicationStoppedEvent
 import ru.vyarus.dropwizard.guice.test.GuiceyAppRule
 import spock.lang.Specification
 
@@ -21,28 +22,35 @@ class ShutdownEventTest extends Specification {
     def "Check shutdown event called"() {
         def rule = new DropwizardAppRule(App) {}
         App.shutdown = null
+        App.stopped = null
 
         when: "start-stop with jetty app"
         rule.apply({} as Statement, null).evaluate()
         then: "shutdown called"
         App.shutdown != null
         App.shutdown
+        App.stopped != null
+        App.stopped
     }
 
     def "Check shutdown event called in lightweight tests"() {
         def rule = new GuiceyAppRule(App, null)
         App.shutdown = null
+        App.stopped = null
 
         when: "start-stop without jetty app"
         rule.apply({} as Statement, null).evaluate()
         then: "shutdown called"
         App.shutdown != null
-        !App.shutdown
+        !App.shutdown // indicate called event, but not started server
+        App.stopped != null
+        App.stopped
     }
 
     static class App extends Application<Configuration> {
 
         static Boolean shutdown
+        static Boolean stopped
 
         @Override
         void initialize(Bootstrap<Configuration> bootstrap) {
@@ -51,6 +59,11 @@ class ShutdownEventTest extends Specification {
                         @Override
                         protected void applicationShutdown(ApplicationShotdownEvent event) {
                             shutdown = event.jettyStarted
+                        }
+
+                        @Override
+                        protected void applicationStopped(ApplicationStoppedEvent event) {
+                            stopped = true
                         }
                     })
                     .build())